Refactored some scripts, prettified loggings etc.

+++ docker/run-script.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-set -euxo pipefail
+set -euo pipefail
 
 # Clean up leftovers before exit
 function cleanup {
-    docker rm $CONTAINER_ID
-    sleep 3
+    if [ "${CONTAINER_ID-}" ]; then
+        docker rm $CONTAINER_ID
+    fi
 }
 trap cleanup EXIT
 
+function on_error {
+    echo "Could not run script '$SCRIPT' inside container" >&2
+    cleanup
+    sleep 5
+    exit 1
+}
+trap on_error ERR
+
+# Check that argument was supplied
 if [[ $# -eq 0 ]] ; then
     echo "No argument supplied"
     exit 1
@@ -29,11 +39,12 @@ DOCKERFILE="${2:-$CURRENT_DIR/Dockerfile.build}"
 IMAGE_TAG="${3:-$REPO_NAME:build}"
 
 # Make sure network is started (used to enable communication by container-name)
-NETWORK=$($DIR/../docker/start-network.sh 2>&1 >/dev/null)
+NETWORK=$($CURRENT_DIR/../docker/start-network.sh 2>&1 >/dev/tty)
 
-IMAGE_ID=$($CURRENT_DIR/build-image.sh $DOCKERFILE $IMAGE_TAG 2>&1 >/dev/null)
+# Make sure image is built
+IMAGE_ID=$($CURRENT_DIR/build-image.sh $DOCKERFILE $IMAGE_TAG 2>&1 >/dev/tty)
 
-echo "Running script '$SCRIPT' inside container..."
+echo -e "\n-- Running script '$SCRIPT' inside container (image (Image: '$IMAGE_TAG')...\n"
 
 # Create build container & compile (create+start instead of run because of issues with logs)
 CONTAINER_ID=$( docker create \
@@ -42,4 +53,7 @@ CONTAINER_ID=$( docker create \
                 --workdir $CONTAINER_WDIR \
                 $IMAGE_ID sh -c "$SCRIPT" \
                 )
+
 docker start -i $CONTAINER_ID
+
+echo -e "\n-- DONE running script '$SCRIPT' inside container (Image: '$IMAGE_TAG')...\n"
